lint: reject circular imports and break the two renderer cycles
Enable Biome's suspicious/noImportCycles at error severity. It is a
built-in project-domain rule, so it sees the whole module graph -- a
GritQL plugin cannot, it only ever sees one file. Default severity is
warn; error is right here because a runtime cycle produces symbols that
are undefined at module init, and a warning nobody has to clear rots.
ignoreTypes stays on: the repo has no verbatimModuleSyntax, so an
`import type` is erased and cuts the cycle at runtime.

It found two cycles, both in the renderer, both the same shape -- the
controller module renders the view, the view reads the controller back
through context:

  App.tsx <-> AppView.tsx
  components/Conversation.tsx <-> components/ConversationView.tsx

Move each context, hook and provider into its own module that borrows
only the controller's *type* back. The HMR boundary the split exists for
is unchanged; only the import direction is.

Cost is ~300ms over 774 files, so `bun run lint` is unaffected.
