+++ CONTRIBUTING.md
@@ -28,6 +28,12 @@ The Biome configuration also loads the repository-owned GritQL rules in
 `tools/biome/anti-slop/rules`. Fix these findings at the domain boundary. Do not suppress a rule or
 replace a concrete contract with a broad dictionary type.
 
+Biome's `suspicious/noImportCycles` rejects circular imports across the whole repository, so runtime
+imports point in one direction. When two modules need each other, move the shared piece into a third
+module both can import — as `app-controller-context.tsx` and `conversation-controller-context.tsx` do
+for their controller contexts. Type-only imports (`import type`) are erased by the compiler and stay
+allowed in both directions.
+
 ## Pull requests
 
 1. Create a branch from `main`.
